vpn-l2tp: store IPsec PSK as a secret
Since nm-l2tp 1.20.0 release (oct. 2021), the default is to store the
IPsec PSK as a secret rather than a setting item.

Store the PSK as a secret and load the legacy setting if present for
automatic migration.

Without this change, if the PSK was stored as a secret, it would be lost
on save.