Add missing ModemManager 1.20 dependency
commit 3e44cf047b3e22349e37c0ec1990346136df24f0 "Manages new cases
introduced in ModemManager 1.20" missed actually raising the minimum
required version.

Which it could not do anyway without adding the missing dependency
in the first place - so far it had been relying on mm-qt to provide
it transiently, but that is stuck on a 1.0 minimum since 2014.

FindModemManager.cmake imported from modemmanager-qt.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>