feat: default enterprise wifi connections to per-user encrypted storage
After the change to make connections global by default, this change
ensures that specifically enterprise wifi connections set up through the
KCM will default to making the connection per-user.

This is more appropriate for enterprise networks where the manner in
which you connect to the network may be identifying for that specific
user, and this being the default for that type of connection is a more
appropriate and secure default.